Removed equals from PlaySession.
This is to allow the system to treat two sessions as distinct even if they have the same game and play time on the same date. Added static comparator method to PlayData to manage times when the date objects may not match when comparing two PlayData objects. (Not 100% sure this is a proper forever idea, but we get different dates after saving / loading the PlayData)

Updated the test for loading data to use the comparator method so the test works propperly now.