Added Aggregator interface and Aggregate class.
Aggregator will structure how new aggregates are calculated and added to the data.

The PlayAggregate class contains a map of games & data, for each game there can be zero or more aggregate fields. Each field includes a type and a value. At this point all values must be doubles (which works with the current data we might aggregate).