Fixed the filters in CLI to behave as expected.
Fixing issue #22, changed the behaviour / design of filters in the CLI. Added 'clear' and 'isEmpty' methods to the filter interface (which were basically already in the implementation classes).

Note: a filter with NOTHING in it will still return an empty Play Data. The CLI now checks if a filter is empty and if it is, skips applying the filter.