Added .equals to PlaySession
Removed because this was causeing some odd behaviour in the CSVSessionManager, but I think the switch to LocalDate has largely fixed this (or just made it worse, but the tests look good). So I added in the equals methods and depricated the weird static compairson method I added earlier. (see #27 and commit 1504d018069e4397116921718fcc355bef4b99f0)