Changed Aggregator to Super Class from Interface.
Seemed like a good change because all of the Aggregator implementors are
strictly aggregators that do nothing else. As such I'm switching to the super
class. It may make sense to switch it back later but this seems like a
`cleaner' implementation at this point.

Also cleaned up a few static anallysis catches.