Merging 2021 Refresh Development (#55)
* Updated .gitignore to ignore .DS_Store files

* (Very old) edit to a comment

* Modified CLI to take preference set as param.

Allows for future where we can split the CLI from the main controls. Will also make it easier to modularize the rest of the system.

* Horrifying everything update.

Most of the things in 2021 Refresh are included in here, including the unification of the persistance managers, the unusable TypedJSONPersistenceManager and the functioning CSVPersistenceManager.

Don't do this at home kids. Or anywhere. Just don't do this.

* Added .csv, .data and .json to gitignore.

I've used all of them as datafiles in the ugly update. Until they're needed for admin needs I'm globaly ignoring them.